Move SquidCursorFactory into the adapter package
